SignTalk/requirements.txt
absl-py==1.1.0
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
asgiref==3.5.2
asttokens==2.0.5
astunparse==1.6.3
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.6.1
autocommand==2.2.1
Automat==20.2.0
backcall==0.2.0
backports.csv==1.0.7
# backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bs4==0.0.1
cachetools==5.2.0
certifi==2022.6.15
cffi==1.15.1
channels==3.0.5
channels-redis==3.4.1
charset-normalizer==2.1.0
cheroot==8.6.0
CherryPy==18.8.0
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
constantly==15.1.0
cryptography==37.0.4
cycler==0.11.0
daphne==3.0.2
datasets==2.4.0
debugpy==1.6.0
decorator==4.4.2
dill==0.3.5.1
Django==4.0.6
django-appconf==1.0.5
django-bootstrap==0.2.4
django-bootstrap4==22.2
django-compressor==4.1
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.13.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
feedparser==6.0.10
filelock==3.8.0
flatbuffers==1.12
fonttools==4.33.3
frozenlist==1.3.1
fsspec==2022.8.2
future==0.18.2
futures==2.2.0
gast==0.4.0
gitdb==4.0.9
GitPython==3.1.27
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
goslate==1.5.4
grpcio==1.47.0
h5py==3.7.0
hiredis==2.0.0
huggingface-hub==0.9.1
hyperlink==21.0.0
idna==3.3
imageio==2.22.0
imageio-ffmpeg==0.4.7
importlib-metadata==4.12.0
importlib-resources==5.9.0
incremental==21.3.0
inflect==6.0.0
ipykernel==6.15.0
ipython==8.4.0
jaraco.classes==3.2.3
jaraco.collections==3.5.2
jaraco.context==4.1.2
jaraco.functools==3.5.2
jaraco.text==3.9.1
jedi==0.18.1
joblib==1.1.0
jupyter-client==7.3.4
jupyter-core==4.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
keybert==0.6.0
keytotext==2.3.2
kiwisolver==1.4.3
libclang==14.0.1
lxml==4.9.1
Markdown==3.3.7
matplotlib==3.5.2
matplotlib-inline==0.1.3
mediapipe==0.8.10
more-itertools==8.14.0
moviepy==1.0.3
msgpack==1.0.4
multidict==6.0.2
multiprocess==0.70.13
mysqlclient==2.1.1
nest-asyncio==1.5.5
nltk==3.7
numpy==1.23.0
oauthlib==3.2.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
parso==0.8.3
pathtools==0.1.2
Pattern==3.6
pdfminer.six==20220524
pickleshare==0.7.5
Pillow==9.1.1
portend==3.1.0
proglog==0.1.10
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.1
pure-eval==0.2.2
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.10.2
pyDeprecate==0.3.2
PyDictionary==2.0.1
Pygments==2.12.0
pyOpenSSL==22.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==0.21.0
pytorch-lightning==1.7.5
pytz==2022.1
pywin32==304
PyYAML==6.0
pyzmq==23.2.0
rcssmin==1.1.0
regex==2022.8.17
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.5.1
rjsmin==1.2.0
rsa==4.8
scikit-learn==1.1.1
scipy==1.8.1
sentence-transformers==2.2.2
sentencepiece==0.1.97
sentry-sdk==1.9.8
service-identity==21.1.0
setproctitle==1.3.2
sgmllib3k==1.0.0
shortuuid==1.0.9
six==1.15.0
sklearn==0.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.3.0
tempora==5.0.2
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tokenizers==0.12.1
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.9.3
torchvision==0.13.1
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
transformers==4.21.3
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.12
wandb==0.13.3
wcwidth==0.2.5
Werkzeug==2.1.2
wrapt==1.12.1
xxhash==3.0.0
yarl==1.8.1
zc.lockfile==2.0
zipp==3.8.0
zope.interface==5.4.0
